Move create/switch action to episode_records controller
