+++ app/services/page_data/episode_records_new_service.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module PageData
+  class EpisodeRecordsNewService < PageData::WorkDetailService
+  end
+end
