+++ app/views/api/internal/page_data/episode_records_new.jb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+render "work_detail", results: @results
