+++ app/views/application/page_params/episode_records_new.jb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+render "application/page_params/work_detail", work: @work
