+++ app/views/episodes/show.html+mobile.slim
@@ -17,7 +17,7 @@
 .c-card.mb-3.py-3
   - if user_signed_in?
     .container
-      = render "records/form", record: @record, url: records_path(current_user.username)
+      = render "records/form", record: @record, url: episode_records_path(@episode)
     hr
   .container
     - if user_signed_in?
