+++ app/views/records/_form.html.slim
@@ -1,7 +1,6 @@
 = form_for record, url: url do |f|
   = render "application/components/error_panel", resource: record
   
-  = f.hidden_field :episode_id
   = hidden_field_tag :page_category, page_category
 
   .mb-2
