+++ config/routes/web.rb
@@ -87,7 +87,7 @@ end
 
 resources :episodes, only: [] do
   resources :items, only: %i(new destroy), controller: :episode_items
-  resources :records, only: [] do
+  resources :records, only: %i(create), controller: "episode_records" do
     post :switch, on: :collection
   end
 end
@@ -123,7 +123,7 @@ scope "@:username", username: /[A-Za-z0-9_]+/ do
   resources :favorite_people, only: %i(index)
   resources :tags, only: %i(show), controller: :user_work_tags, as: :user_work_tag
 
-  resources :records, only: %i(create show edit update destroy) do
+  resources :records, only: %i(show edit update destroy) do
     resources :comments, only: %i(create)
   end
